Adding Haystack and Solr (with a *tiny* bit of setup instructions)
This commit squashes a few commits that represented false starts and errors of various sorts.

This puts enough indexing in place to allow a sort of "full text" search of each lexical entry from the dirt-simple Haystack basic search view.

Next steps will include formatting the output of the search and making it possible to narrow the search to particular fields (especially "headword"). The former takes priority for purposes of getting a demo up.